Extract `gemm::tests` module into separate file
Split out the large `tests` module into its own file.
